# swagger.yaml
openapi: 3.0.4
info:
  title: API de Agenda de Ajustes Pós-Cirúrgicos
  description: |
    Sistema completo para gerenciamento de agenda de ajustes de produtos pós-cirúrgicos.

    ## Características
    - Autenticação JWT com controle de permissões
    - Cadastro de pessoas com diferentes papéis
    - Gestão de agendamentos com regras de negócio
    - Upload de fotos de ajustes
    - Sistema de logs para auditoria
    - Notificações via WhatsApp
    - Relatórios e dashboard

    ## Tipos de Usuário
    - **Administrador**: Acesso total ao sistema
    - **Secretária**: Cadastro de pacientes e agendamentos
    - **Costureira**: Registro de comparecimento e upload de fotos
    - **Médico**: Visualização de relatórios
    - **Paciente**: Recebe notificações (sem acesso ao sistema)

    ## Autenticação
    A maioria dos endpoints requer autenticação via JWT. 
    Obtenha um token através do endpoint `/api/auth/login` e inclua-o no header:
    ```
    Authorization: Bearer {seu_token}
    ```

  version: 1.0.0
  contact:
    name: Ramon David Souto
    email: rdssi@hotmail.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.sistema.com
    description: Servidor de produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do login
  schemas:
...
    Error:
      type: object
      properties:
        sucesso:
          type: boolean
          example: false
        mensagem:
          type: string
          example: Erro na operação
        erro:
          type: string
          example: Detalhes do erro

    # Schema de Sucesso
    Success:
      type: object
      properties:
        sucesso:
          type: boolean
          example: true
        mensagem:
          type: string
          example: Operação realizada com sucesso

    # Schema de Login
...
    PessoaInput:
      type: object
      required:
        - id_cidade
        - ind_tipo_pessoa
        - nom_completo
        - num_cpf
        - num_rg
        - dta_nascimento
        - des_logradouro
        - num_endereco
        - num_cep
        - nom_bairro
        - num_celular_1
        - des_email_1
      properties:
        id_cidade:
          type: integer
          description: ID da cidade
          example: 1
        ind_status:
          type: string
          enum: [ativo, inativo]
          default: ativo
          example: ativo
        ind_tipo_pessoa:
          type: string
          enum: [medico, secretaria, costureira, paciente, administrador]
          description: Tipo de pessoa/usuário
          example: paciente
        nom_completo:
          type: string
          maxLength: 255
          description: Nome completo
          example: João da Silva Santos
        num_cpf:
          type: string
          minLength: 11
          maxLength: 11
          pattern: '^\d{11}$'
          description: CPF sem formatação
          example: "12345678901"
        num_rg:
          type: string
          maxLength: 20
          description: RG
          example: "1234567"
        dta_nascimento:
          type: string
          format: date
          description: Data de nascimento
          example: "1990-05-15"
        des_logradouro:
          type: string
          maxLength: 255
          description: Logradouro (rua, avenida, etc)
          example: Rua das Flores
        num_endereco:
          type: string
          maxLength: 10
          description: Número do endereço
          example: "123"
        des_complemento:
          type: string
          maxLength: 100
          nullable: true
          description: Complemento do endereço
          example: Apartamento 201
        num_cep:
          type: string
          minLength: 8
          maxLength: 8
          pattern: '^\d{8}$'
          description: CEP sem formatação
          example: "74000000"
        nom_bairro:
          type: string
          maxLength: 100
          description: Nome do bairro/setor
          example: Centro
        num_celular_1:
          type: string
          maxLength: 15
          description: Celular principal
          example: "62999999999"
        ind_whatsapp_1:
          type: boolean
          default: true
          description: Indica se celular 1 tem WhatsApp
          example: true
        num_celular_2:
          type: string
          maxLength: 15
          nullable: true
          description: Celular secundário
          example: "62988888888"
        ind_whatsapp_2:
          type: boolean
          default: true
          description: Indica se celular 2 tem WhatsApp
          example: false
        des_email_1:
          type: string
          format: email
          maxLength: 255
          description: Email principal
          example: joao@email.com
        des_email_2:
          type: string
          format: email
          maxLength: 255
          nullable: true
          description: Email secundário
          example: joao2@email.com
        des_senha:
          type: string
          minLength: 6
          description: Senha (obrigatória para secretaria, costureira, admin e medico)
          example: senha123
    Pessoa:
      allOf:
        - $ref: "#/components/schemas/PessoaInput"
        - type: object
          properties:
            id:
              type: integer
              example: 1
            nom_cidade:
              type: string
              example: Goiânia
            nom_estado:
              type: string
              example: Goiás
            sgl_estado:
              type: string
              example: GO
            created_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
...
    Agendamento:
      type: object
      properties:
        id:
          type: integer
          example: 1
        dta_agendamento:
          type: string
          format: date-time
          example: "2024-12-31T14:00:00Z"
        ind_status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado, falta]
          example: agendado
        ind_paciente_compareceu:
          type: boolean
          nullable: true
          example: null
        des_observacoes_geral:
          type: string
          nullable: true
          example: Paciente necessita de ajuste urgente
        des_observacoes_costureira:
          type: string
          nullable: true
          example: Ajuste realizado conforme solicitado
        paciente_id:
          type: integer
          example: 5
        paciente_nome:
          type: string
          example: João da Silva Santos
        paciente_cpf:
          type: string
          example: "12345678901"
        paciente_celular:
          type: string
          example: "62999999999"
        paciente_email:
          type: string
          example: joao@email.com
        costureira_id:
          type: integer
          example: 3
        costureira_nome:
          type: string
          example: Ana Costureira Santos
        secretaria_id:
          type: integer
          example: 2
        secretaria_nome:
          type: string
          example: Maria Secretária Silva
        produtos:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              id_produto_grade:
                type: integer
                example: 10
              nom_produto:
                type: string
                example: Blusa Manga Longa
              ref_produto:
                type: string
                example: BML-001
              num_sku:
                type: string
                example: BML-001-P-AZ
              cod_tamanho:
                type: string
                example: P
              nom_cor_estampa:
                type: string
                example: Azul
              des_ajuste_produto:
                type: string
                example: Ajustar manga esquerda
              ind_status:
                type: string
                enum: [pendente, producao, concluido, cancelado]
                example: pendente
              fotos:
                type: array
                items:
                  $ref: "#/components/schemas/Foto"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ComparecimentoInput:
      type: object
      required:
        - ind_paciente_compareceu
      properties:
        ind_paciente_compareceu:
          type: boolean
          description: Indica se o paciente compareceu
          example: true
        des_observacoes_costureira:
          type: string
          nullable: true
          description: Observações da costureira sobre o atendimento
          example: Paciente compareceu e ajuste foi realizado com sucesso

    # Schema de Foto
    Foto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_produto_agendamento:
          type: integer
          example: 1
        des_caminho_arquivo:
          type: string
          example: uploads/ajustes/foto-1234567890.jpg
        nom_original:
          type: string
          example: foto_ajuste.jpg
        num_tamanho_bytes:
          type: integer
          example: 2048576
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
  responses:
    UnauthorizedError:
      description: Token de autenticação ausente ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Não autorizado
            erro: Token não fornecido ou inválido
    ForbiddenError:
      description: Usuário não tem permissão para executar esta ação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Acesso negado
            erro: Você não tem permissão para executar esta ação
paths:
...
  /api/auth/verificar:
    get:
      tags:
        - Autenticação
      summary: Verificar token JWT
      description: Verifica se o token JWT é válido e retorna informações do usuário
      operationId: verificarToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  mensagem:
                    type: string
                    example: Token válido
                  dados:
                    type: object
                    properties:
                      usuario:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          cpf:
                            type: string
                            example: "00000000000"
                          tipo:
                            type: string
                            example: administrador
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/perfil:
    get:
      tags:
        - Autenticação
      summary: Obter perfil do usuário logado
      description: Retorna os dados completos do usuário autenticado
      operationId: obterPerfil
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  dados:
                    $ref: "#/components/schemas/Pessoa"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/alterar-senha:
    put:
      tags:
        - Autenticação
      summary: Alterar senha do próprio usuário
      description: Permite que o usuário altere sua própria senha
      operationId: alterarSenha
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - senha_atual
                - nova_senha
                - confirmar_senha
              properties:
                senha_atual:
                  type: string
                  description: Senha atual do usuário
                  example: senha123
                nova_senha:
                  type: string
                  minLength: 6
                  description: Nova senha (mínimo 6 caracteres)
                  example: novaSenha456
                confirmar_senha:
                  type: string
                  description: Confirmação da nova senha
                  example: novaSenha456
      responses:
        "200":
          description: Senha alterada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
              example:
                sucesso: true
                mensagem: Senha alterada com sucesso
        "400":
          description: Erro na validação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                senhasNaoCoincidem:
                  summary: Senhas não coincidem
                  value:
                    sucesso: false
                    mensagem: Erro de validação
                    erro: Nova senha e confirmação não coincidem
                senhaAtualIncorreta:
                  summary: Senha atual incorreta
                  value:
                    sucesso: false
                    mensagem: Senha atual incorreta
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/auth/resetar-senha:
    post:
      tags:
        - Autenticação
      summary: Resetar senha de um usuário (apenas admin)
      description: Permite que um administrador resete a senha de qualquer usuário
      operationId: resetarSenha
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - num_cpf
                - nova_senha
              properties:
                num_cpf:
                  type: string
                  pattern: '^\d{11}$'
                  description: CPF do usuário que terá a senha resetada
                  example: "12345678901"
                nova_senha:
                  type: string
                  minLength: 6
                  description: Nova senha para o usuário
                  example: novaSenha789
      responses:
        "200":
          description: Senha resetada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
              example:
                sucesso: true
                mensagem: Senha resetada com sucesso
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ==================== PESSOAS ====================
...
  /api/agendamentos/{id}/comparecimento:
    post:
      tags:
        - Agendamentos
      summary: Registrar comparecimento do paciente
      description: |
        Registra se o paciente compareceu ao agendamento e adiciona observações da costureira.

        **Regras:**
        - Apenas a costureira responsável pelo agendamento pode registrar
        - Altera o status para 'concluido' ou 'falta'

        **Permissões:**
        - Costureira: Pode registrar seus próprios agendamentos
        - Administrador: Pode registrar qualquer agendamento
      operationId: registrarComparecimento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do agendamento
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparecimentoInput"
      responses:
        "200":
          description: Comparecimento registrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  mensagem:
                    type: string
                    example: Comparecimento registrado com sucesso
                  dados:
                    $ref: "#/components/schemas/Agendamento"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Agendamento não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...